refactor: reorganize instance metadata procedure
Closes #2207
