+++ controllers/cluster_controller_test.go
@@ -22,17 +22,13 @@ import (
 
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
-	schemeBuilder "github.com/cloudnative-pg/cloudnative-pg/internal/scheme"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/postgres"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/reconciler/persistentvolumeclaim"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/utils"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -269,454 +265,6 @@ var _ = Describe("Updating target primary", func() {
 				Namespace: cluster.Namespace,
 			}, &expectedPod)
 			Expect(err).ToNot(HaveOccurred())
-		})
-	})
-})
-
-var _ = Describe("object metadata test", func() {
-	makeReconciler := func(pods []corev1.Pod) *ClusterReconciler {
-		builder := fake.NewClientBuilder().
-			WithScheme(schemeBuilder.BuildWithAllKnownScheme())
-
-		for idx := range pods {
-			builder = builder.WithObjects(&pods[idx])
-		}
-
-		return &ClusterReconciler{Client: builder.Build()}
-	}
-
-	getPod := func(re *ClusterReconciler, name string) *corev1.Pod {
-		pod := &corev1.Pod{}
-		err := re.Client.Get(context.Background(), types.NamespacedName{Name: name}, pod)
-		Expect(err).ToNot(HaveOccurred())
-		return pod
-	}
-
-	Context("updateRoleLabelsOnPods", func() {
-		It("Should update the role labels correctly", func() {
-			cluster := &apiv1.Cluster{
-				Status: apiv1.ClusterStatus{
-					CurrentPrimary: "primaryPod",
-				},
-			}
-
-			pods := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:   "primaryPod",
-						Labels: map[string]string{},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:   "replicaPod",
-						Labels: map[string]string{},
-					},
-				},
-			}
-
-			re := makeReconciler(pods)
-			err := re.updateRoleLabelsOnPods(context.Background(), cluster, corev1.PodList{Items: pods})
-			Expect(err).ToNot(HaveOccurred())
-
-			primaryPod := getPod(re, pods[0].Name)
-			Expect(primaryPod.Labels[specs.ClusterRoleLabelName]).To(Equal(specs.ClusterRoleLabelPrimary))
-
-			replicaPod := getPod(re, pods[1].Name)
-			Expect(replicaPod.Labels[specs.ClusterRoleLabelName]).To(Equal(specs.ClusterRoleLabelReplica))
-		})
-
-		It("Should update the role labels when the primary and the replica switch roles", func() {
-			cluster := &apiv1.Cluster{
-				Status: apiv1.ClusterStatus{
-					CurrentPrimary: "newPrimaryPod",
-				},
-			}
-
-			pods := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "newPrimaryPod",
-						Labels: map[string]string{
-							specs.ClusterRoleLabelName: specs.ClusterRoleLabelReplica,
-						},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "newReplicaPod",
-						Labels: map[string]string{
-							specs.ClusterRoleLabelName: specs.ClusterRoleLabelPrimary,
-						},
-					},
-				},
-			}
-
-			re := makeReconciler(pods)
-			err := re.updateRoleLabelsOnPods(context.Background(), cluster, corev1.PodList{Items: pods})
-			Expect(err).ToNot(HaveOccurred())
-
-			newPrimaryPod := getPod(re, pods[0].Name)
-			Expect(newPrimaryPod.Labels[specs.ClusterRoleLabelName]).To(Equal(specs.ClusterRoleLabelPrimary))
-
-			newReplicaPod := getPod(re, pods[1].Name)
-			Expect(newReplicaPod.Labels[specs.ClusterRoleLabelName]).To(Equal(specs.ClusterRoleLabelReplica))
-		})
-
-		It("Should not perform role reconciliation when there is no current primary", func() {
-			cluster := &apiv1.Cluster{}
-
-			pods := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:   "oldPrimaryPod",
-						Labels: map[string]string{specs.ClusterRoleLabelName: specs.ClusterRoleLabelPrimary},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:   "oldReplicaPod",
-						Labels: map[string]string{specs.ClusterRoleLabelName: specs.ClusterRoleLabelReplica},
-					},
-				},
-			}
-
-			re := makeReconciler(pods)
-			err := re.updateRoleLabelsOnPods(context.Background(), cluster, corev1.PodList{Items: pods})
-			Expect(err).ToNot(HaveOccurred())
-
-			// Labels of pods should not have been updated
-			for _, pod := range pods {
-				updatedPod := getPod(re, pod.Name)
-				Expect(updatedPod.Labels[specs.ClusterRoleLabelName]).To(Equal(pod.Labels[specs.ClusterRoleLabelName]))
-			}
-		})
-	})
-
-	Context("updateOperatorLabelsOnInstances", func() {
-		const instanceName = "instance1"
-		It("Should create labels if the instance has no labels", func() {
-			instances := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: instanceName,
-					},
-				},
-			}
-
-			re := makeReconciler(instances)
-			err := re.updateOperatorLabelsOnInstances(context.Background(), corev1.PodList{Items: instances})
-			Expect(err).ToNot(HaveOccurred())
-
-			updatedInstance := getPod(re, instances[0].Name)
-			Expect(updatedInstance.Labels[utils.InstanceNameLabelName]).To(Equal(instances[0].Name))
-			Expect(updatedInstance.Labels[utils.PodRoleLabelName]).To(Equal(string(utils.PodRoleInstance)))
-		})
-
-		It("Should not update labels if the instance already has the correct labels", func() {
-			instances := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: instanceName,
-						Labels: map[string]string{
-							utils.InstanceNameLabelName: instanceName,
-							utils.PodRoleLabelName:      string(utils.PodRoleInstance),
-						},
-					},
-				},
-			}
-
-			re := makeReconciler(instances)
-			err := re.updateOperatorLabelsOnInstances(context.Background(), corev1.PodList{Items: instances})
-			Expect(err).ToNot(HaveOccurred())
-
-			// Labels of instances should not have been updated
-			for _, instance := range instances {
-				updatedInstance := getPod(re, instance.Name)
-				Expect(updatedInstance.Labels).To(Equal(instance.Labels))
-			}
-		})
-
-		It("Should update name label if it's incorrect", func() {
-			instances := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: instanceName,
-						Labels: map[string]string{
-							utils.InstanceNameLabelName: "incorrectName",
-							utils.PodRoleLabelName:      string(utils.PodRoleInstance),
-						},
-					},
-				},
-			}
-
-			re := makeReconciler(instances)
-			err := re.updateOperatorLabelsOnInstances(context.Background(), corev1.PodList{Items: instances})
-			Expect(err).ToNot(HaveOccurred())
-
-			updatedInstance := getPod(re, instances[0].Name)
-			Expect(updatedInstance.Labels[utils.InstanceNameLabelName]).To(Equal(instanceName))
-		})
-
-		It("Should update role label if it's incorrect", func() {
-			instances := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: instanceName,
-						Labels: map[string]string{
-							utils.InstanceNameLabelName: instanceName,
-							utils.PodRoleLabelName:      "incorrectRole",
-						},
-					},
-				},
-			}
-
-			re := makeReconciler(instances)
-			err := re.updateOperatorLabelsOnInstances(context.Background(), corev1.PodList{Items: instances})
-			Expect(err).ToNot(HaveOccurred())
-
-			updatedInstance := getPod(re, instances[0].Name)
-			Expect(updatedInstance.Labels[utils.PodRoleLabelName]).To(Equal(string(utils.PodRoleInstance)))
-		})
-
-		It("Should correctly update labels for multiple instances", func() {
-			instances := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: instanceName,
-						Labels: map[string]string{
-							utils.InstanceNameLabelName: "incorrectName1",
-							utils.PodRoleLabelName:      string(utils.PodRoleInstance),
-						},
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "instance2",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "instance3",
-						Labels: map[string]string{
-							utils.InstanceNameLabelName: "instance3",
-							utils.PodRoleLabelName:      "incorrectRole",
-						},
-					},
-				},
-			}
-
-			re := makeReconciler(instances)
-			err := re.updateOperatorLabelsOnInstances(context.Background(), corev1.PodList{Items: instances})
-			Expect(err).ToNot(HaveOccurred())
-
-			for _, instance := range instances {
-				updatedInstance := getPod(re, instance.Name)
-				Expect(updatedInstance.Labels[utils.InstanceNameLabelName]).To(Equal(updatedInstance.Name))
-				Expect(updatedInstance.Labels[utils.PodRoleLabelName]).To(Equal(string(utils.PodRoleInstance)))
-			}
-		})
-	})
-
-	Context("updateClusterLabelsOnPods", func() {
-		const (
-			labelKey      = "label1"
-			labelValue    = "value1"
-			labelKeyTwo   = "label2"
-			labelValueTwo = "value2"
-		)
-
-		It("Should correctly add missing labels from cluster to pods", func() {
-			cluster := &apiv1.Cluster{
-				Spec: apiv1.ClusterSpec{
-					InheritedMetadata: &apiv1.EmbeddedObjectMetadata{
-						Labels: map[string]string{
-							labelKey:    labelValue,
-							labelKeyTwo: labelValueTwo,
-						},
-					},
-				},
-			}
-
-			pods := corev1.PodList{
-				Items: []corev1.Pod{
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "pod1",
-							Labels: map[string]string{
-								labelKey: labelValue,
-							},
-						},
-					},
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "pod2",
-						},
-					},
-				},
-			}
-
-			re := makeReconciler(pods.Items)
-			err := re.updateClusterLabelsOnPods(context.Background(), cluster, pods)
-			Expect(err).ToNot(HaveOccurred())
-
-			for _, pod := range pods.Items {
-				updatedPod := getPod(re, pod.Name)
-				Expect(updatedPod.Labels).To(Equal(cluster.GetFixedInheritedLabels()))
-			}
-		})
-
-		It("Should not change labels if they already match the cluster's", func() {
-			cluster := &apiv1.Cluster{
-				Spec: apiv1.ClusterSpec{
-					InheritedMetadata: &apiv1.EmbeddedObjectMetadata{
-						Labels: map[string]string{
-							labelKey:    labelValue,
-							labelKeyTwo: labelValueTwo,
-						},
-					},
-				},
-			}
-			pods := corev1.PodList{
-				Items: []corev1.Pod{
-					{
-						ObjectMeta: metav1.ObjectMeta{
-							Name: "pod1",
-							Labels: map[string]string{
-								labelKey:    labelValue,
-								labelKeyTwo: labelValueTwo,
-							},
-						},
-					},
-				},
-			}
-
-			Expect(cluster.Spec.InheritedMetadata.Labels).To(Equal(cluster.GetFixedInheritedLabels()))
-
-			re := makeReconciler(pods.Items)
-			err := re.updateClusterLabelsOnPods(context.Background(), cluster, pods)
-			Expect(err).ToNot(HaveOccurred())
-
-			updatedPod := getPod(re, pods.Items[0].Name)
-			Expect(updatedPod.Labels).To(Equal(cluster.GetFixedInheritedLabels()))
-		})
-
-		It("Should correctly handle the case of no fixed inherited labels from the cluster", func() {
-			cluster := &apiv1.Cluster{
-				Spec: apiv1.ClusterSpec{
-					InheritedMetadata: &apiv1.EmbeddedObjectMetadata{},
-				},
-			}
-			pods := corev1.PodList{
-				Items: []corev1.Pod{{ObjectMeta: metav1.ObjectMeta{Name: "pod1"}}},
-			}
-			re := makeReconciler(pods.Items)
-			err := re.updateClusterLabelsOnPods(context.Background(), cluster, pods)
-			Expect(err).ToNot(HaveOccurred())
-
-			updatedPod := getPod(re, pods.Items[0].Name)
-			Expect(updatedPod.Labels).To(BeEmpty())
-		})
-	})
-
-	Context("updateClusterAnnotationsOnPods", func() {
-		const key = "annotation1"
-		const value = "value1"
-
-		It("Should correctly add missing annotations from cluster to pods", func() {
-			pods := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pod1",
-					},
-				},
-			}
-			cluster := &apiv1.Cluster{
-				Spec: apiv1.ClusterSpec{
-					InheritedMetadata: &apiv1.EmbeddedObjectMetadata{
-						Annotations: map[string]string{key: value},
-					},
-				},
-			}
-			re := makeReconciler(pods)
-			err := re.updateClusterAnnotationsOnPods(context.Background(), cluster, corev1.PodList{Items: pods})
-			Expect(err).NotTo(HaveOccurred())
-
-			updatedPod := getPod(re, pods[0].Name)
-			Expect(updatedPod.Annotations[key]).To(Equal(value))
-		})
-
-		It("Should not change annotations if they already match the cluster's", func() {
-			pods := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:        "pod1",
-						Annotations: map[string]string{key: value},
-					},
-				},
-			}
-			cluster := &apiv1.Cluster{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: apiv1.ClusterSpec{
-					InheritedMetadata: &apiv1.EmbeddedObjectMetadata{
-						Annotations: map[string]string{key: value},
-					},
-				},
-			}
-
-			Expect(cluster.Spec.InheritedMetadata.Annotations).To(Equal(cluster.GetFixedInheritedAnnotations()))
-
-			re := makeReconciler(pods)
-			err := re.updateClusterAnnotationsOnPods(context.Background(), cluster, corev1.PodList{Items: pods})
-			Expect(err).NotTo(HaveOccurred())
-
-			updatedPod := getPod(re, pods[0].Name)
-			Expect(updatedPod.Annotations).To(HaveLen(1))
-			Expect(updatedPod.Annotations[key]).To(Equal(value))
-		})
-
-		It("Should correctly add AppArmor annotations if present in the cluster's annotations", func() {
-			const (
-				key   = "container.apparmor.security.beta.kubernetes.io/pod"
-				value = "runtime/default"
-			)
-
-			pods := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pod1",
-					},
-				},
-			}
-
-			cluster := &apiv1.Cluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{key: value},
-				},
-			}
-			re := makeReconciler(pods)
-			err := re.updateClusterAnnotationsOnPods(context.Background(), cluster, corev1.PodList{Items: pods})
-			Expect(err).NotTo(HaveOccurred())
-
-			updatedPod := getPod(re, pods[0].Name)
-			Expect(updatedPod.Annotations[key]).To(Equal(value))
-		})
-
-		It("Should correctly handle the case of no fixed inherited annotations from the cluster", func() {
-			pods := []corev1.Pod{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "pod1",
-					},
-				},
-			}
-			cluster := &apiv1.Cluster{}
-			re := makeReconciler(pods)
-			err := re.updateClusterAnnotationsOnPods(context.Background(), cluster, corev1.PodList{Items: pods})
-			Expect(err).NotTo(HaveOccurred())
-
-			updatedPod := getPod(re, pods[0].Name)
-			Expect(len(updatedPod.Annotations)).To(Equal(0))
 		})
 	})
 })
