+++ controllers/replicas.go
@@ -26,7 +26,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	apiv1 "github.com/cloudnative-pg/cloudnative-pg/api/v1"
-	"github.com/cloudnative-pg/cloudnative-pg/internal/configuration"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/log"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/postgres"
 	"github.com/cloudnative-pg/cloudnative-pg/pkg/specs"
@@ -369,178 +368,6 @@ func (r *ClusterReconciler) evaluateFailoverDelay(
 	delay := time.Duration(failOverDelay) * time.Second
 	if delay > primaryFailingSince {
 		return ErrWaitingOnFailOverDelay
-	}
-
-	return nil
-}
-
-// updateClusterAnnotationsOnPods we check if we need to add or modify existing annotations specified in the cluster but
-// not existing in the pods. We do not support the case of removed annotations from the cluster resource.
-func (r *ClusterReconciler) updateClusterAnnotationsOnPods(
-	ctx context.Context,
-	cluster *apiv1.Cluster,
-	pods corev1.PodList,
-) error {
-	contextLogger := log.FromContext(ctx)
-
-	for i := range pods.Items {
-		pod := &pods.Items[i]
-
-		// if all the required annotations are already set and with the correct value,
-		// we proceed to the next item
-		if utils.IsAnnotationSubset(pod.Annotations, cluster.Annotations, cluster.GetFixedInheritedAnnotations(),
-			configuration.Current) &&
-			utils.IsAnnotationAppArmorPresentInObject(&pod.ObjectMeta, cluster.Annotations) {
-			contextLogger.Debug(
-				"Skipping cluster annotations reconciliation, because they are already present on pod",
-				"pod", pod.Name,
-				"podAnnotations", pod.Annotations,
-				"clusterAnnotations", cluster.Annotations,
-			)
-			continue
-		}
-
-		// otherwise, we add the modified/new annotations to the pod
-		patch := client.MergeFrom(pod.DeepCopy())
-		utils.InheritAnnotations(&pod.ObjectMeta, cluster.Annotations,
-			cluster.GetFixedInheritedAnnotations(), configuration.Current)
-		if utils.IsAnnotationAppArmorPresent(cluster.Annotations) {
-			utils.AnnotateAppArmor(&pod.ObjectMeta, cluster.Annotations)
-		}
-		contextLogger.Info("Updating cluster annotations on pod", "pod", pod.Name)
-		if err := r.Patch(ctx, pod, patch); err != nil {
-			return err
-		}
-		continue
-	}
-
-	return nil
-}
-
-// updateClusterAnnotationsOnPods we check if we need to add or modify existing labels specified in the cluster but
-// not existing in the pods. We do not support the case of removed labels from the cluster resource.
-func (r *ClusterReconciler) updateClusterLabelsOnPods(
-	ctx context.Context,
-	cluster *apiv1.Cluster,
-	pods corev1.PodList,
-) error {
-	contextLogger := log.FromContext(ctx)
-
-	for i := range pods.Items {
-		pod := &pods.Items[i]
-
-		// if all the required labels are already set and with the correct value,
-		// we proceed to the next item
-		if utils.IsLabelSubset(pod.Labels, cluster.Labels, cluster.GetFixedInheritedLabels(),
-			configuration.Current) {
-			contextLogger.Debug(
-				"Skipping cluster label reconciliation, because they are already present on pod",
-				"pod", pod.Name,
-				"podLabels", pod.Labels,
-				"clusterLabels", cluster.Labels,
-			)
-			continue
-		}
-
-		// otherwise, we add the modified/new labels to the pod
-		patch := client.MergeFrom(pod.DeepCopy())
-		utils.InheritLabels(&pod.ObjectMeta, cluster.Labels, cluster.GetFixedInheritedLabels(), configuration.Current)
-
-		contextLogger.Info("Updating cluster labels on pod", "pod", pod.Name)
-		if err := r.Patch(ctx, pod, patch); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// Make sure that only the currentPrimary has the label forward write traffic to him
-func (r *ClusterReconciler) updateRoleLabelsOnPods(
-	ctx context.Context,
-	cluster *apiv1.Cluster,
-	pods corev1.PodList,
-) error {
-	contextLogger := log.FromContext(ctx)
-
-	// No current primary, no work to do
-	if cluster.Status.CurrentPrimary == "" {
-		return nil
-	}
-
-	primaryFound := false
-	for idx := range pods.Items {
-		pod := &pods.Items[idx]
-
-		if !utils.IsPodActive(*pod) {
-			contextLogger.Info("Ignoring not active Pod during label update",
-				"pod", pod.Name, "status", pod.Status)
-			continue
-		}
-
-		podRole, hasRole := pod.ObjectMeta.Labels[specs.ClusterRoleLabelName]
-
-		switch {
-		case pod.Name == cluster.Status.CurrentPrimary:
-			primaryFound = true
-
-			if !hasRole || podRole != specs.ClusterRoleLabelPrimary {
-				contextLogger.Info("Setting primary label", "pod", pod.Name)
-				patch := client.MergeFrom(pod.DeepCopy())
-				pod.Labels[specs.ClusterRoleLabelName] = specs.ClusterRoleLabelPrimary
-				if err := r.Patch(ctx, pod, patch); err != nil {
-					return err
-				}
-			}
-
-		default:
-			if !hasRole || podRole != specs.ClusterRoleLabelReplica {
-				contextLogger.Info("Setting replica label", "pod", pod.Name)
-				patch := client.MergeFrom(pod.DeepCopy())
-				pod.Labels[specs.ClusterRoleLabelName] = specs.ClusterRoleLabelReplica
-				if err := r.Patch(ctx, pod, patch); err != nil {
-					return err
-				}
-			}
-		}
-	}
-
-	if !primaryFound {
-		contextLogger.Info("No primary instance found for this cluster")
-	}
-
-	return nil
-}
-
-// updateOperatorLabelsOnInstances ensures that the instances have the correct labels
-func (r *ClusterReconciler) updateOperatorLabelsOnInstances(
-	ctx context.Context,
-	instances corev1.PodList,
-) error {
-	for i := range instances.Items {
-		instance := &instances.Items[i]
-		if instance.Labels == nil {
-			instance.Labels = map[string]string{}
-		}
-		var modified bool
-		origInstance := instance.DeepCopy()
-
-		if instance.Labels[utils.InstanceNameLabelName] != instance.Name {
-			instance.Labels[utils.InstanceNameLabelName] = instance.Name
-			modified = true
-		}
-
-		if instance.Labels[utils.PodRoleLabelName] != string(utils.PodRoleInstance) {
-			instance.Labels[utils.PodRoleLabelName] = string(utils.PodRoleInstance)
-			modified = true
-		}
-		if !modified {
-			continue
-		}
-
-		if err := r.Client.Patch(ctx, instance, client.MergeFrom(origInstance)); err != nil {
-			return err
-		}
 	}
 
 	return nil
